+++ tests/testthat/test-classify_noise.R
@@ -1,3 +1,5 @@
+# ===== IVF ====
+
 test_that("classify noise with ivf works",
 {
   f <- system.file("extdata", "Topography.las", package="lasR")
@@ -19,6 +21,8 @@ test_that("classify noise with ivf works using unorder map",
   expect_equal(ans$npoints_per_class, c(`1` = 60721, `2` = 8053, `9` = 3835, `18` = 794))
 })
 
+#  ===== SOR =====
+
 test_that("classify noise with sor works",
 {
   f <- system.file("extdata", "Topography.las", package="lasR")
@@ -38,8 +42,29 @@ test_that("classify noise with sor works",
 
 test_that("classify noise with sor fails if k < 2",
 {
-  f <- system.file("extdata", "Topography.las", package="lasR")
-  class = classify_with_sor(k = 1)
-  expect_error(exec(class, f), "less than 2-nearest neighbors")
+  expect_error(classify_with_sor(k = 1), "less than 2-nearest neighbors")
 })
+
+# ===== IPF ====
+
+test_that("classify noise with ipf works",
+{
+  f <- system.file("extdata", "Topography.las", package="lasR")
+  class = classify_with_ipf(r = 4, n = 0L)
+  ans = exec(class+summarise(), f)
+
+  expect_equal(ans$npoints_per_class, c(`1` = 60267, `2` = 7954, `9` = 3889, `18` = 1293))
+
+  class = classify_with_ipf(r = 4, n = 1L)
+  ans = exec(class+summarise(), f)
+
+  expect_equal(ans$npoints_per_class, c(`1` = 56907, `2` = 7365, `9` = 3860, `18` = 5271))
+})
+
+
+test_that("classify noise with ipf fails if r < 0",
+{
+  expect_error(classify_with_ipf(r = -1), "radius must be positive")
+})
+